Add dependency for bundler.
It has been changed to use the bundler for the gems dependency management. The Gemfile in the root
directory is the description file for that information.

Once you follow the updated instructions in INSTALL.txt, the depend gems will be installed under
vendor/gems/. config.ru has also been modified to refer the gems in local vendor folder.