add RunInitializer Rack middleware to run initializers at the first time HTTP request.
PhusionPassenger and Unicorn use the technique for pre-load origin process. This makes
an origin process to load config.ru and then spawns new worker process from that
one. It can save the machine resource but the user has to run initialization code
carefully.

The problem is that the initialization codes only run in the initial process if
config.ru runs initialization codes directly. The spawned worker process inherits
the artifacts as result of those initialization but they are not valid in the spawned
processes: DB connections or TCP sockets.

Those web servers provides the special hooks to solve this problem. However, it needs the
changes for each and brings complexity to the initializers.

This Rack middleware allows to run initialization codes inside of the spawned process not in
the origin process.