change transaction isolation level to READ COMMITTED.
For MySQL Cluster, the engine does not support REPEATABLE READ
isolation level. So we set the READ COMMITTED isolation level
exclusively for each MySQL connection and also on other database
engines.

Note on AUTOCOMMIT mode change. the mode pushed back to AUTOCOMMIT=1.
When we sends isolation level change query, it was rejected because
the transaction is opened always due to AUTOCOMMIT=0.